complete tokenizer refactor
