+++ torchtune/models/gemma/_model_builders.py
@@ -8,7 +8,7 @@ from typing import List
 from torchtune.models.gemma._component_builders import gemma, lora_gemma
 from torchtune.models.gemma.transformer import GemmaTransformerDecoder
 
-from torchtune.modules.tokenizers import SentencePieceTokenizer
+from torchtune.modules.tokenizers import SentencePieceEncoding
 from torchtune.modules.peft import LORA_ATTN_MODULES
 
 from functools import partial
@@ -41,10 +41,8 @@ def gemma_2b() -> GemmaTransformerDecoder:
     )
 
 
-def gemma_tokenizer(path: str) -> SentencePieceTokenizer:
-    tokenizer = SentencePieceTokenizer(path)
-    tokenizer.pad_id = 0
-    return tokenizer
+def gemma_tokenizer(path: str) -> GemmaTokenizer:
+    return GemmaTokenizer(path)
 
 
 def lora_gemma_2b(
