+++ torchtune/models/phi3/_model_builders.py
@@ -1,7 +1,7 @@
 from typing import List
 
 from torchtune.models.phi3._component_builders import phi3, lora_phi3
-from torchtune.models.phi3._sentencepiece import Phi3MiniSentencePieceTokenizer
+from torchtune.models.phi3._sentencepiece import Phi3MiniSentencePieceEncoding
 
 from torchtune.modules import TransformerDecoder
 from torchtune.modules.peft import LORA_ATTN_MODULES
@@ -40,7 +40,7 @@ def phi3_mini() -> TransformerDecoder:
         norm_eps=1e-5,
     )
 
-def phi3_mini_tokenizer(path: str) -> Phi3MiniSentencePieceTokenizer:
+def phi3_mini_tokenizer(path: str) -> Phi3MiniTokenizer:
     """Phi-3 Mini tokenizer.
     Ref: https://huggingface.co/microsoft/Phi-3-mini-4k-instruct/blob/main/tokenizer_config.json
 
@@ -58,10 +58,9 @@ def phi3_mini_tokenizer(path: str) -> Phi3MiniSentencePieceTokenizer:
         See https://huggingface.co/microsoft/Phi-3-mini-4k-instruct/discussions/51 for more details.
 
     Returns:
-        Phi3MiniSentencePieceTokenizer: Instantiation of the SPM tokenizer.
+        Phi3MiniSentencePieceEncoding: Instantiation of the SPM tokenizer.
     """
-    tokenizer = Phi3MiniSentencePieceTokenizer(path)
-    return tokenizer
+    return Phi3MiniTokenizer(path)
 
 
 def lora_phi3_mini(
