+++ torchtune/models/phi3/_tokenizer.py
@@ -8,9 +8,10 @@ from typing import List, Optional, Tuple
 
 from sentencepiece import SentencePieceProcessor
 from torchtune.data import Message, truncate
+from torchtune.modules.tokenizers import Tokenizer
 
 
-class Phi3MiniSentencePieceTokenizer:
+class Phi3MiniTokenizer(Tokenizer):
     """A wrapper around SentencePieceProcessor.
 
     Args:
@@ -18,8 +19,8 @@ class Phi3MiniSentencePieceTokenizer:
 
     Example:
         # Accepts only non-batched input for now
-        >>> tokenizer = SentencePieceTokenizer("/path/to/spm_model")
-        >>> tokenized_text = SentencePieceTokenizer.encode("Hello world!", add_bos=True, add_eos=True)
+        >>> tokenizer = SentencePieceEncoding("/path/to/spm_model")
+        >>> tokenized_text = SentencePieceEncoding.encode("Hello world!", add_bos=True, add_eos=True)
         >>> print(tokenized_text)
         [1, 31587, 29644, 102, 2]
     """
@@ -28,31 +29,23 @@ class Phi3MiniSentencePieceTokenizer:
         self,
         path: str,
     ):
-        spm_model = SentencePieceProcessor()
-        spm_model.load(path)
-        self.spm_model = spm_model
+        self._spm_model = SentencePieceEncoding(path)
 
-        self.special_tokens = {
-            "<|endoftext|>": 32000,
-            "<|assistant|>": 32001,
-            "<|placeholder1|>": 32002,
-            "<|placeholder2|>": 32003,
-            "<|placeholder3|>": 32004,
-            "<|placeholder4|>": 32005,
-            "<|system|>": 32006,
-            "<|end|>": 32007,
-            "<|placeholder5|>": 32008,
-            "<|placeholder6|>": 32009,
-            "<|user|>": 32010,
-        }
+        self.special_tokens = self._get_special_tokens()
 
-        self.vocab_size = spm_model.vocab_size()
-        self.bos_id = spm_model.bos_id()
         self.eos_id = self.special_tokens["<|endoftext|>"]
         self.pad_id = self.special_tokens["<|endoftext|>"]
 
         # During generation, stop when eos_id is encountered
         self.stop_tokens = [self.eos_id]
+
+    @property
+    def eos_id(self):
+        return self._spm_model.eos_id
+
+    @property
+    def vocab_size(self):
+        return self._spm_model.vocab_size
 
     def encode(
         self,
@@ -60,45 +53,10 @@ class Phi3MiniSentencePieceTokenizer:
         add_bos: bool = True,
         add_eos: bool = True,
         trim_leading_whitespace: bool = False,
-        prefix: Optional[str] = None,
     ) -> List[int]:
-        """Encode text into token IDs.
-
-        Args:
-            text (str): The input text to be encoded, unbatched.
-            add_bos (bool): Whether to prepend BOS to the input, defaults to True.
-            add_eos (bool): Whether to append EOS to the input, defaults to True.
-            trim_leading_whitespace (bool): Whether to trim leading whitespace from
-                underlying sentencepiece tokenization. Sentencepiece normally prepends
-                whitespace to any tokenized text, which can cause differences where
-                encode(s1) + encode(s2) != encode(s1 + s2) due to leading whitespace
-                added to s2. Default: False
-            prefix (Optional[str]): Optional string to encode for trimming leading
-                whitespaces. Used only if trim_leading_whitespace=True. Default: None
-        Returns:
-            List[int]: The encoded token IDs.
-        """
-        if trim_leading_whitespace:
-            # Can define our own custom prefix depending on vocab if needed
-            if not hasattr(self, "prefix"):
-                self.prefix = prefix or "\n"
-                self.encoded_prefix = self.spm_model.encode(
-                    self.prefix, add_bos=False, add_eos=False
-                )
-            start_idx = len(self.encoded_prefix) + int(add_bos)
-            return self.spm_model.encode(
-                self.prefix + text,
-                add_bos=add_bos,
-                add_eos=add_eos,
-                out_type=int,
-            )[start_idx:]
-        else:
-            return self.spm_model.encode(
-                text,
-                add_bos=add_bos,
-                add_eos=add_eos,
-                out_type=int,
-            )
+        return self._spm_model.encode(
+            text, add_bos=add_bos, add_eos=add_eos, trim_leading_whitespace=trim_leading_whitespace
+        )
 
     def decode(self, ids: List[int]) -> str:
         """Decode token IDs to strings.
@@ -117,7 +75,7 @@ class Phi3MiniSentencePieceTokenizer:
                 continue
             else:
                 ids_for_decode.append(token_id)
-        return self.spm_model.decode(ids_for_decode)
+        return self._spm_model.decode(ids_for_decode)
 
     def tokenize_messages(
         self,
@@ -131,7 +89,7 @@ class Phi3MiniSentencePieceTokenizer:
         returning a list of tokens and a list of masks.
 
         Example:
-            >>> tokenizer = SentencePieceTokenizer(tokenizer_path)
+            >>> tokenizer = SentencePieceEncoding(tokenizer_path)
             >>> messages = [
                 Message(role="system", content="system message\n", masked=True),
                 Message(role="user", content="user prompt\n", masked=True),
